==> notebooks/1_hot100_scraping.ipynb <==
Import Libraries

import requests
from bs4 import BeautifulSoup
import pandas as pd

Define the URL and get the response

url = "https://www.billboard.com/charts/hot-100/"
response = requests.get(url)
soup = BeautifulSoup(response.content, "html.parser")


Scrape data for position, track, artist and image url

song_titles = []
artists = []
positions = []

chart_items = soup.select('ul.o-chart-results-list-row')

for item in chart_items:
    title = item.select_one('h3#title-of-a-story').text.strip() # extract title
    artist = item.select_one('span.c-label.a-no-trucate').text.strip() # extract artist
    
    song_titles.append(title)
    artists.append(artist)

hot100_df = pd.DataFrame({
    'title': song_titles,
    'artist': artists
})

display(hot100_df)

Save to csv in the data folder

hot100_df.to_csv('../data/1_hot100_cleaned.csv', index=False)


**Make sure notebooks have:**

* A quick description at the top of what is going to be done
* Headers with numbers on every section
* Describe any decision made in a sentence
* Make sure to run the cells and save the notebook with the output. outputs should be readable, not to long and make sure no errors are shown.
