==> notebooks/3_combined_cleaned_df.ipynb <==
Libraries and Data Imports

import pandas as pd

# ... existing imports and data loading ...
df_hot100 = pd.read_csv('../data/1_hot100_cleaned.csv')
df_2m = pd.read_csv('../data/2_million_songs_cleaned.csv')

Merging Dataframes, removing duplicates, adding isHot column

df_combined = pd.concat([df_hot100, df_2m], ignore_index=True)

# Add new column indicating if song is from Hot 100
df_combined['isHot'] = 'No'
df_combined.loc[:len(df_hot100)-1, 'isHot'] = 'Yes'

# Remove duplicates, keeping Hot 100 versions when available
df_combined = df_combined.sort_values('isHot', ascending=False).drop_duplicates(subset='title', keep='first')


Export to CSV

df_combined.to_csv('../data/3_combined_cleaned_df.csv', index=False)


**Make sure notebooks have:**

* A quick description at the top of what is going to be done
* Headers with numbers on every section
* Describe any decision made in a sentence
* Make sure to run the cells and save the notebook with the output. outputs should be readable, not to long and make sure no errors are shown.
